DP-38339-Add-unique-ID-to-login-dropdown-component (#1967)
* Update button-dropdown.twig

adds a conditional id attribute in button-dropdown.twig, so we can set gtm-loginto on the "Log In To" button for tracking.

* Create DP-38339.yml

adds a change log.

* Update button.twig

This ensures that the id attribute is only added to the button if it is explicitly provided.

* Update page-header.twig

Pass the `button` object as-is to `@molecules/button-dropdown.twig` and `@atoms/01-buttons/button.twig`, allowing OpenMass to provide the `id` if needed.

* Revert "Update page-header.twig"

This reverts commit 1861c6146fb60101ea9f78892fcf38ff28825d59.

* Revert "Update button-dropdown.twig"

This reverts commit 203728e05bf21616ab3bd37ee642c6d091c056cb.